Move hackathon code to /app/archive
Move AI4E hackathon showcase, judging API, and related components
to archive directory. These are no longer actively used but preserved
for reference.

Moved:
- /app/hacks → /app/archive/hacks/main
- /app/ai4e-hacks → /app/archive/hacks/ai4e
- /app/api/judge → /app/archive/api/judge
- /app/components/ProjectCard.tsx → /app/archive/components/
- /app/lib/projects.ts → /app/archive/lib/
