+++ app/archive/README.md
@@ -1,0 +1,28 @@
+# Archive
+
+This directory contains archived code that is no longer actively used but preserved for reference.
+
+## Contents
+
+### `/hacks` - AI for Epistemics Hackathon (March 2025)
+
+Hackathon showcase pages for the AI for Epistemics hackathon event.
+
+- `/hacks/main` - Main hackathon showcase page (`/archive/hacks/main`)
+- `/hacks/ai4e` - Alias page (`/archive/hacks/ai4e`)
+
+### `/api/judge` - Claude Judging Endpoint
+
+API endpoint that used Claude to provide feedback on hackathon projects.
+
+### `/components/ProjectCard.tsx`
+
+React component for displaying hackathon project cards with Claude judging functionality.
+
+### `/lib/projects.ts`
+
+Utility for fetching hackathon projects from Airtable (uses separate AI4E Airtable base).
+
+## Note
+
+These routes are still accessible at `/archive/hacks/main` etc. if needed, but are not linked from the main site. The associated Airtable base and API keys may no longer be active.
