+++ app/archive/hacks/ai4e/page.tsx
@@ -1,4 +1,4 @@
-import Hacks from '../hacks/page'
+import Hacks from '../main/page'
 
 export default function AI4EHacks() {
   return <Hacks />
